+++ test/inspector_test.dart
@@ -10,11 +10,11 @@ void main() {
         (tester) async {
       final List<String> logs = <String>[];
 
-      void onValueUpdated<T>(T old, T value) {
+      void onAction<T>(T old, T value, String action) {
         logs.add('$old => $value');
       }
 
-      final MaestroInspector inspector = MaestroInspector(onValueUpdated);
+      final MaestroInspector inspector = MaestroInspector(onAction);
       final DefaultComposer composer = DefaultComposer();
 
       await tester.pumpWidget(
@@ -36,6 +36,37 @@ void main() {
 
       composer.store('b');
       expect(logs, <String>['1 => 2', 'a => b']);
+    });
+
+    testWidgets('method is called with action name', (tester) async {
+      final List<String> logs = <String>[];
+
+      void onAction<T>(T old, T value, String action) {
+        logs.add('$action: $old => $value');
+      }
+
+      final MaestroInspector inspector = MaestroInspector(onAction);
+      final DefaultComposer composer = DefaultComposer();
+
+      await tester.pumpWidget(
+        Maestros(
+          [
+            Maestro(inspector),
+            const Maestro(1),
+            const Maestro('a'),
+            Maestro(composer),
+          ],
+          child: const SizedBox(),
+        ),
+      );
+
+      expect(logs, isEmpty);
+
+      composer.store(2, 'IntStorage');
+      expect(logs, <String>['IntStorage: 1 => 2']);
+
+      composer.store('b', 'StringStorage');
+      expect(logs, <String>['IntStorage: 1 => 2', 'StringStorage: a => b']);
     });
   });
 }