+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## [0.0.1]
+## [0.1.0]
+### Added
+- An `Object` parameter to `write` and `readAndWrite` methods, in order to indicate what the action is.
 
-* Initial Open Source release.
+### Changed
+- The method in `MaestroInspector` from `onValueUpdated` to `onAction` along with the signature method.
+
+## [0.0.1]
+- Initial Open Source release.